mgmt: mcumgr: transport: shell: fix compilation error
#63967 removed the struct declaration required by
`smp_shell_input_timeout_handler`, update the function
accordingly so that it compiles.
